+++ src/App.tsx
@@ -19,7 +19,9 @@ function App() {
             enable={process.env.NODE_ENV === 'development'}
             // Breakpoints used to determine the current responsive behavior.
             // By default, Tailwind CSS breakpoints are used.
-            breakpoints="tailwind" // Use preset 'tailwind', 'bootstrap', 'bootstrap4', 'bootstrap5', 'foundation', 'bulma', 'mui' or provide custom breakpoints
+            // You can use one of the presets: 'tailwind', 'bootstrap', 'bootstrap4', 'bootstrap5', 'foundation', 'bulma', 'mui'
+            // or provide custom breakpoints as an object.
+            breakpoints="tailwind"
             // Example of custom breakpoints:
             // breakpoints={{
             //   XS: [0, 639],
@@ -40,11 +42,13 @@ function App() {
             showNextBreakpoint={true}
             // If false, hides the close button in the overlay (default: true)
             showCloseButton={true}
-            // Transparency level of the overlay. A value between 0 (fully transparent) and 1 (fully opaque)
-            transparency={0.95} // (default: 1)
-            // Theme can be one of: 'light' | 'dark' | 'scheme' | 'class' | CustomTheme (by default 'scheme')
+            // Transparency level of the overlay.
+            // A value between 0 (fully transparent) and 1 (fully opaque). Default is 1
+            transparency={0.95}
+            // Theme for the overlay.
+            // Can be one of: 'light', 'dark', 'scheme', 'class', 'green', 'indigo', 'orange', or a custom theme object.
+            // Default is 'scheme'
             theme="dark"
-
             // If you want a fully customized color scheme,
             // pass a theme object instead of a preset value:
             // theme={{
@@ -55,6 +59,14 @@ function App() {
             //   closeButtonColor: '#2e7400', // Color for the close button (if showCloseButton=true)
             //   fontFamily: 'Arial, sans-serif', // Font family for all text in the overlay
             // }}
+
+            // Custom container styles.
+            // These styles are applied to the outer container of the overlay and can be used to override the default positioning and z-index.
+            // By default, z-index is set to 1000.
+            containerStyles={{ zIndex: 1000, bottom: 16, right: 16 }}
+            // Custom overlay styles.
+            // These styles are applied directly to the overlay element, allowing further customization (e.g., font size, padding).
+            // overlayStyles={{ fontSize: '10px' }}
           />
         )}
       </Suspense>